fix problem 1: update deploy-rs; ensure use it; nixpkgs follows it
The important one is updating deploy-rs. making nixpkgs follow
deploy-rs/nixpkgs is incidental, but is an easy way to avoid how much to
update nixpkgs!

We also make sure the `deploy-rs` that is doing the deploy is the one
pinned in the flake, to avoid this out-of-sync error in the future